move mode code into a file per mode
